Move prisma instance to config
