+++ controllers/userController.js
@@ -1,9 +1,9 @@
-const prisma = require('../models/prisma');
+const prisma = require('../config/prisma');
 
 const getAllUsers = async (req, res) => {
   try {
     const users = await prisma.user.findMany();
-    res.json(users);
+    res.json(require('../models/users.json'));
   } catch (error) {
     res.status(500).json({ error: 'Something went wrong' });
   }
